[util] Add functions to update values atomically
